preprocessing-worker/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pillow>=10.0.0
opencv-python-headless>=4.9.0
imagehash>=4.3.0
albumentations>=1.4.0
boto3>=1.34.0
httpx>=0.27.0
pydantic>=2.0.0
numpy>=1.26.0
